+++ mayan/apps/metadata/managers.py
@@ -1,5 +1,6 @@
 from __future__ import unicode_literals
 
+from django.apps import apps
 from django.db import models
 
 
@@ -19,4 +20,17 @@ class MetadataTypeManager(models.Manager):
             pk__in=document_type.metadata.values_list(
                 'metadata_type', flat=True
             )
+        )
+
+
+class DocumentTypeMetadataTypeManager(models.Manager):
+    def get_metadata_types_for(self, document_type):
+        DocumentType = apps.get_model(
+            app_label='metadata', model_name='MetadataType'
+        )
+
+        return DocumentType.objects.filter(
+            pk__in=self.filter(
+                document_type=document_type
+            ).values_list('metadata_type__pk', flat=True)
         )